requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
pytest==8.3.3
pytest-cov==5.0.0
psycopg2-binary==2.9.9
email-validator==2.2.0
